More flexibility in accepting DateTime identifiers
Allows `chrono::DateTime` alongside `DateTime`
Fixes #53

The real issue here is that the migration type resolution is relying
on the proc macro, which sees only the AST and not the type system. So
when it sees a type identifier, it sees the path segments that are in
the AST, not the actual type, meaning that `chrono::DateTime` and
`DateTime` (after `use chrono::DateTime`) are not the same! The latter
was getting accepted and the former is not.

This PR is a bandaid to discard the path information and just look at
the last segment (hoping it was imported from chrono!) to make this
slightly more flexible.

This type of issue is the biggest reason I want to move migration
generation from proc-macro time to something triggered by the `butane`
cli command.